requirements.txt
absl-py==0.6.1
alabaster==0.7.12
amqp==2.3.2
aniso8601==3.0.2
apturl==0.5.2
argcomplete==1.8.1
argh==0.26.2
asn1crypto==0.24.0
astor==0.7.1
astroid==2.0.4
attrs==18.2.0
Automat==0.7.0
Babel==2.6.0
backcall==0.1.0
bcrypt==3.1.4
billiard==3.5.0.4
bleach==3.0.2
blessings==1.6
bpython==0.17.1
Brlapi==0.6.6
celery==4.2.1
certifi==2019.3.9
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
command-not-found==0.3
confusable-homoglyphs==3.2.0
constantly==15.1.0
cryptography==2.3.1
cssselect==1.0.3
cupshelpers==1.0
curtsies==0.2.12
cycler==0.10.0
decorator==4.3.0
defer==1.0.6
defusedxml==0.5.0
distro-info==0.18
Django==2.1.1
django-appconf==1.0.2
django-celery==3.1.17
django-countries==5.3.2
django-elasticsearch-dsl==0.5.0
django-ordered-model==2.1.0
django-payfast==0.3
-e git://github.com/cypreess/django-plans.git@5cf0c429efd31482aca43ca3a3049c3b2b07153e#egg=django_plans
django-queued-storage==0.8.0
django-registration==3.0
django-rest-auth==0.9.3
django-storages==1.7.1
django-user-accounts==2.0.3
django-webpack-loader==0.6.0
djangorestframework==3.8.2
docutils==0.14
elasticsearch==6.3.1
elasticsearch-dsl==6.1.0
entrypoints==0.2.3
Fabric==1.10.2
Fabric3==1.14.post1
fake-useragent==0.1.11
falcon==1.3.0
Flask==1.0.2
Flask-Babel==0.11.2
Flask-GraphQL==2.0.0
fleep==1.0.1
gast==0.2.0
geographiclib==1.49
geopy==1.17.0
gevent==1.3.6
graphene==2.1.3
graphene-sqlalchemy==2.1.0
graphql-core==2.1
graphql-relay==0.4.5
graphql-server-core==1.1.1
greenlet==0.4.15
grpcio==1.17.0
gunicorn==19.9.0
h5py==2.8.0
httplib2==0.9.2
hyperlink==18.0.0
idna==2.7
imagesize==1.1.0
incremental==17.5.0
invoke==1.2.0
ipaddress==1.0.22
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
iso8601==0.1.12
isort==4.3.4
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kazam==1.4.5
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
kombu==4.2.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
louis==3.5.0
lxml==4.2.5
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
netifaces==0.10.4
notebook==5.7.0
numpy==1.16.1
numpydoc==0.8.0
oauth==1.0.1
olefile==0.45.1
opencv-python==4.0.0.21
packaging==18.0
pandas==0.23.4
pandocfilters==1.4.2
paramiko==2.4.2
parsel==1.5.0
parso==0.3.1
passlib==1.7.1
pathtools==0.1.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.0.0
pipenv==2018.11.26
playsound==1.2.2
prometheus-client==0.4.2
promise==2.2.1
prompt-toolkit==2.0.7
protobuf==3.6.1
psutil==5.4.6
psycopg2==2.7.7
psycopg2-binary==2.7.7
ptyprocess==0.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycairo==1.16.3
pycodestyle==2.4.0
pycountry==18.2.23
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
PyDispatcher==2.0.5
pyfiglet==0.8.post0
pyflakes==2.0.0
pygame==1.9.4
Pygments==2.2.0
pygobject==3.26.1
PyHamcrest==1.9.0
PyJWT==1.5.3
pylint==2.1.1
pymacaroons==0.13.0
PyNaCl==1.2.1
pynput==1.4
pyOpenSSL==18.0.0
pyparsing==2.3.1
PyPDF2==1.26.0
PyQt5==5.9.2
pyRFC3339==1.0
python-apt==1.6.3
python-dateutil==2.8.0
python-debian==0.1.32
python-logstash==0.4.6
python-magic==0.4.15
python-mimeparse==1.6.0
python-stdnum==1.9
python-xlib==0.23
pytz==2018.7
pyxdg==0.25
PyYAML==3.12
pyzmq==17.1.2
QtAwesome==0.5.3
qtconsole==4.4.2
QtPy==1.5.2
queuelib==1.5.0
reportlab==3.4.0
requests==2.19.1
requests-unixsocket==0.1.5
rope==0.11.0
Rx==1.6.1
scipy==1.1.0
Scrapy==1.5.1
scrapy-fake-useragent==1.1.0
scrapy-splash==0.7.2
seaborn==0.9.0
SecretStorage==2.3.1
Send2Trash==1.5.0
service-identity==17.0.0
simplejson==3.13.2
singledispatch==3.4.0.3
sip==4.19.8
six==1.12.0
snowballstemmer==1.2.1
Sphinx==1.8.2
sphinxcontrib-websupport==1.1.0
spyder==3.3.1
spyder-kernels==0.2.6
SQLAlchemy==1.2.17
SQLAlchemy-Utils==0.32.21
suds-jurko==0.6
system-service==0.3
systemd-python==234
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
Twisted==18.9.0
typed-ast==1.1.0
typing==3.6.6
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.23
usb-creator==0.3.3
vatnumber==1.2
vboxapi==1.0
vine==1.1.4
virtualenv==16.4.3
virtualenv-clone==0.5.1
w3lib==1.19.0
wadllib==1.3.2
Wand==0.4.4
watchdog==0.8.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.10.11
xkit==0.0.0
zope.interface==4.5.0
